Altera retorno do tipo do método get da propriedade numero_casa

+++ src/main/java/desafiocriptografiajuliocesar/http/payload/CriptografiaJulioCesarPayload.java
@@ -35,7 +35,7 @@ public class CriptografiaJulioCesarPayload {
         this.resumoCriptografado = resumoCriptografado;
     }
 
-    public int getNumeroCasas() {
+    public short getNumeroCasas() {
         return numeroCasas;
     }
